+++ rpcs3/Emu/RSX/VK/VKGSRender.cpp
@@ -7,6 +7,7 @@
 #include "VKRenderPass.h"
 #include "VKResourceManager.h"
 #include "VKCommandStream.h"
+#include "vkutils/buffer_view.h"
 #include "Emu/RSX/rsx_methods.h"
 #include "Emu/Memory/vm_locking.h"
