+++ rpcs3/Emu/RSX/VK/VKHelpers.h
@@ -1125,53 +1125,6 @@ namespace vk
 		VkDevice m_device;
 	};
 
-	struct buffer_view
-	{
-		VkBufferView value;
-		VkBufferViewCreateInfo info = {};
-
-		buffer_view(VkDevice dev, VkBuffer buffer, VkFormat format, VkDeviceSize offset, VkDeviceSize size)
-			: m_device(dev)
-		{
-			info.buffer = buffer;
-			info.format = format;
-			info.offset = offset;
-			info.range = size;
-			info.sType = VK_STRUCTURE_TYPE_BUFFER_VIEW_CREATE_INFO;
-			CHECK_RESULT(vkCreateBufferView(m_device, &info, nullptr, &value));
-		}
-
-		~buffer_view()
-		{
-			vkDestroyBufferView(m_device, value, nullptr);
-		}
-
-		buffer_view(const buffer_view&) = delete;
-		buffer_view(buffer_view&&) = delete;
-
-		bool in_range(u32 address, u32 size, u32& offset) const
-		{
-			if (address < info.offset)
-				return false;
-
-			const u32 _offset = address - static_cast<u32>(info.offset);
-			if (info.range < _offset)
-				return false;
-
-			const auto remaining = info.range - _offset;
-			if (size <= remaining)
-			{
-				offset = _offset;
-				return true;
-			}
-
-			return false;
-		}
-
-	private:
-		VkDevice m_device;
-	};
-
 	class event
 	{
 		VkDevice m_device = VK_NULL_HANDLE;
