+++ rpcs3/Emu/RSX/VK/vkutils/buffer_view.h
@@ -1,0 +1,23 @@
+#pragma once
+
+#include "../VulkanAPI.h"
+
+namespace vk
+{
+	struct buffer_view
+	{
+		VkBufferView value;
+		VkBufferViewCreateInfo info = {};
+
+		buffer_view(VkDevice dev, VkBuffer buffer, VkFormat format, VkDeviceSize offset, VkDeviceSize size);
+		~buffer_view();
+
+		buffer_view(const buffer_view&) = delete;
+		buffer_view(buffer_view&&)      = delete;
+
+		bool in_range(u32 address, u32 size, u32& offset) const;
+
+	private:
+		VkDevice m_device;
+	};
+}
