+++ rpcs3/VKGSRender.vcxproj
@@ -40,6 +40,7 @@
     <ClInclude Include="Emu\RSX\VK\VKShaderInterpreter.h" />
     <ClInclude Include="Emu\RSX\VK\VKTextOut.h" />
     <ClInclude Include="Emu\RSX\VK\VKTextureCache.h" />
+    <ClInclude Include="Emu\RSX\VK\vkutils\buffer_view.h" />
     <ClInclude Include="Emu\RSX\VK\vkutils\chip_class.h" />
     <ClInclude Include="Emu\RSX\VK\vkutils\fence.h" />
     <ClInclude Include="Emu\RSX\VK\vkutils\graphics_pipeline_state.h" />
@@ -73,6 +74,7 @@
     <ClCompile Include="Emu\RSX\VK\VKResourceManager.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKShaderInterpreter.cpp" />
     <ClCompile Include="Emu\RSX\VK\VKTexture.cpp" />
+    <ClCompile Include="Emu\RSX\VK\vkutils\buffer_view.cpp" />
     <ClCompile Include="Emu\RSX\VK\vkutils\chip_class.cpp" />
     <ClCompile Include="Emu\RSX\VK\vkutils\fence.cpp" />
     <ClCompile Include="Emu\RSX\VK\vkutils\memory_block.cpp" />
