+++ rpcs3/VKGSRender.vcxproj.filters
@@ -44,6 +44,9 @@
     <ClCompile Include="Emu\RSX\VK\vkutils\shared.cpp">
       <Filter>vkutils</Filter>
     </ClCompile>
+    <ClCompile Include="Emu\RSX\VK\vkutils\buffer_view.cpp">
+      <Filter>vkutils</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Emu\RSX\VK\VKCommonDecompiler.h" />
@@ -100,6 +103,9 @@
       <Filter>vkutils</Filter>
     </ClInclude>
     <ClInclude Include="Emu\RSX\VK\vkutils\supported_extensions.h">
+      <Filter>vkutils</Filter>
+    </ClInclude>
+    <ClInclude Include="Emu\RSX\VK\vkutils\buffer_view.h">
       <Filter>vkutils</Filter>
     </ClInclude>
   </ItemGroup>
